% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{GenerateAtable}
\alias{GenerateAtable}
\title{Generate tables that store different design elements}
\usage{
GenerateAtable(nClass, K, M, order)
}
\arguments{
\item{nClass}{Number of Attribute Classes}

\item{K}{Number of Attributes}

\item{M}{Number of Responses}

\item{order}{Highest interaction order to consider.
Default model-specified \code{k}.}
}
\value{
Return a \code{list} containing the table caches for different parameters
}
\description{
Each table provides a "cache" of pre-computed values.
}
\details{
This is \strong{an internal function} briefly used to simulate data and, thus, has
been exported into \emph{R} as well as documented. \strong{Output from this function can
change in future versions.}
}
