% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ohoegdm-wrapper.R
\name{ohoegdm}
\alias{ohoegdm}
\title{Ordinal Higher-Order General Diagnostic Model under the
Exploratory Framework (OHOEGDM)}
\usage{
ohoegdm(
  y,
  k,
  m = 2,
  order = k,
  sdMH = 0.4,
  burnin = 1000L,
  chain_length = 10000L,
  l0 = c(1, rep(100, sum(choose(k, seq_len(order))))),
  l1 = c(1, rep(1, sum(choose(k, seq_len(order))))),
  m0 = 0,
  bq = 1
)
}
\arguments{
\item{y}{Item Matrix}

\item{k}{Dimension to estimate for Q matrix}

\item{m}{Number of Item Categories. Default is \code{2} matching the binary case.}

\item{order}{Highest interaction order to consider. Default model-specified \code{k}.}

\item{sdMH}{Metropolis-Hastings standard deviation tuning parameter.}

\item{burnin}{Amount of Draws to Burn}

\item{chain_length}{Number of Iterations for chain.}

\item{l0, l1, m0, bq}{Additional tuning parameters.}
}
\value{
A \code{egdm_delta} object containing three named lists:

\strong{\code{estimates}}
\itemize{
\item \code{thetas}: Average theta coefficients
\item \code{betas}: Average beta coefficients
\item \code{class}: Average class membership
\item \code{pi}: Average attribute class probability.
\item \code{omega}: Average omega
}

\strong{\code{chain}}
\itemize{
\item \code{thetas}: theta coefficients iterations
\item \code{betas}:  beta coefficients iterations
\item \code{class}:  class membership iterations
\item \code{omega}:  omega iterations
}

\strong{\code{details}}
\itemize{
\item \code{n}: Number of Subjects
\item \code{j}: Number of Items
\item \code{k}: Number of Traits
\item \code{order}: Highest interaction order to consider. Default model-specified \code{k}.
\item \code{sdMH}: Metropolis-Hastings standard deviation tuning parameter.
\item \code{l0}: Spike parameter
\item \code{l1}: Slab parameter
\item \code{m0}, \code{bq}: Additional tuning parameters
\item \code{burnin}: Number of Iterations to discard
\item \code{chain_length}: Number of Iterations to keep
}
}
\description{
Performs the Gibbs sampling routine for a higher-order EGDM.
}
\details{
The \strong{\code{estimates}} list contains the mean information from the sampling
procedure. Meanwhile, the \strong{\code{chain}} list contains full MCMC values. Lastly,
the \strong{\code{details}} list provides information regarding the estimation call.
}
\examples{
\dontrun{
## Configuration
N = 100  # Number of Subjects
J = 20   # Number of Items
K = 3    # Number of Attributes

## TBA
}
}
