% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ohoegdm-package.R
\docType{package}
\name{ohoegdm-package}
\alias{ohoegdm-package}
\alias{_PACKAGE}
\title{ohoegdm: Ordinal Higher-Order Exploratory General Diagnostic Model for Polytomous Data}
\description{
Perform a Bayesian estimation of the ordinal exploratory Higher-order General Diagnostic Model (OHOEGDM) for Polytomous Data described by Culpepper, S. A. and Balamuta, J. J. (In Press) <doi:10.1080/00273171.2021.1985949>.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/tmsalab/ohoegdm}
  \item \url{https://tmsalab.github.io/ohoegdm/}
  \item Report bugs at \url{https://github.com/tmsalab/ohoegdm/issues}
}

}
\author{
\strong{Maintainer}: James Joseph Balamuta \email{balamut2@illinois.edu} (\href{https://orcid.org/0000-0003-2826-8458}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Steven Andrew Culpepper \email{sculpepp@illinois.edu} (\href{https://orcid.org/0000-0003-4226-6176}{ORCID}) [copyright holder]
}

}
\keyword{internal}
